Renamed module back for drop-in compatibility
