+++ setup.py
@@ -4,9 +4,9 @@ SESMailer2
 
 from setuptools import setup, find_packages
 
-import ses_mailer2
+import ses_mailer
 
-PACKAGE = ses_mailer2
+PACKAGE = ses_mailer
 
 setup(
     name=PACKAGE.__NAME__,
@@ -18,7 +18,7 @@ setup(
     long_description=PACKAGE.__doc__,
     url='http://github.com/MarSoft/ses-mailer-2/',
     download_url='http://github.com/MarSoft/ses-mailer-2/tarball/master',
-    py_modules=['ses_mailer2'],
+    py_modules=['ses_mailer'],
     include_package_data=True,
     install_requires=[
         "boto",
